Implemented a baseAbsList for ListView and GridView. Also added a parameter to customize the main layout used.
